Completed structure for Module
