+++ iOS Platform Assessment/Modules/Browser/Fuel/FuelTypeSelector.swift
